Update README.md
The final project for CS2401 is a single large project which is to be submitted in three stages and therefore counts as three projects. For this project we will be implementing a game of OTHELLO, with the final product being a game that can play an intelligent game of OTHELLO against a human opponent.

This game must be derived from the files I have given you (a game class written by the author of a textbook we used to use for this course). The files are game.h and game.cc. I have also included a file called colors.h which was created by a former student who has given us permission to use it. It allows you to adjust the colors of the terminal screen during a text-based console or ssh session. I have altered the play function in the game class, by commenting out some of the code, so that it will work for the first phase of the project. By the end we will return to the original game class with only a couple of little alterations.